Change free text annotation font size to float (#1000)
* Change free text annotation font size to float

* Move free text annotation general parameters to AnnotationsConfig

+++ Zotero/Scenes/Detail/PDF/Models/PDFDocumentAnnotation.swift
@@ -21,7 +21,7 @@ struct PDFDocumentAnnotation {
     let color: String
     let comment: String
     let text: String?
-    var fontSize: UInt?
+    var fontSize: CGFloat?
     var rotation: UInt?
     let sortIndex: String
     let dateModified: Date
